make storage types private
