+++ contracts/axelar-gas-service/src/lib.rs
@@ -1,8 +1,8 @@
 #![no_std]
 
 mod event;
+mod storage_types;
 
-pub mod storage_types;
 pub mod contract;
 pub mod error;
