+++ contracts/axelar-operators/src/contract.rs
@@ -131,3 +131,29 @@ impl AxelarOperators {
         Ok(res)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use axelar_soroban_std::assert_some;
+    use soroban_sdk::testutils::Address as _;
+    use soroban_sdk::{Address, Env};
+
+    use super::{AxelarOperators, AxelarOperatorsClient, DataKey};
+
+    #[test]
+    fn test_initialize() {
+        let env = Env::default();
+        let contract_id = env.register_contract(None, AxelarOperators);
+        let client = AxelarOperatorsClient::new(&env, &contract_id);
+        let user = Address::generate(&env);
+        client.initialize(&user);
+
+        assert_some!(env.as_contract(&contract_id, || {
+            env.storage()
+                .instance()
+                .get::<DataKey, bool>(&DataKey::Initialized)
+        }));
+
+        assert_eq!(client.owner(), user);
+    }
+}
