+++ contracts/axelar-operators/src/lib.rs
@@ -1,9 +1,9 @@
 #![no_std]
 
 mod event;
+mod storage_types;
 
 pub mod contract;
 pub mod error;
-pub mod storage_types;
 
 pub use contract::AxelarOperatorsClient;
